Разделил js файлы на модули. Добавил предполагаемые модули.
